Refactor tests to match new js layout
